# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ДомПульс API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/houses/{house_id}/contacts": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Contacts",
        "operationId": "contacts_api_v1_houses__house_id__contacts_get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        },
        "security": [
          {
            "MAXLaunchData": []
          }
        ]
      }
    },
    "/api/v1/houses/{house_id}/apartments": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Apartments",
        "operationId": "apartments_api_v1_houses__house_id__apartments_get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/houses/{house_id}/issues": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Issues",
        "operationId": "issues_api_v1_houses__house_id__issues_get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resident"
        ],
        "summary": "Report",
        "operationId": "report_api_v1_houses__house_id__issues_post",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssue"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{issue_id}": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Detail",
        "operationId": "detail_api_v1_issues__issue_id__get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/houses/{house_id}/issues/with-photos": {
      "post": {
        "tags": [
          "resident"
        ],
        "summary": "Report With Photos",
        "operationId": "report_with_photos_api_v1_houses__house_id__issues_with_photos_post",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "data",
                  "photos"
                ],
                "properties": {
                  "data": {
                    "type": "string",
                    "description": "JSON-encoded CreateIssue"
                  },
                  "photos": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 10,
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{issue_id}/photos/{photo_id}": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Resident Photo",
        "operationId": "resident_photo_api_v1_issues__issue_id__photos__photo_id__get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          },
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApartmentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApartmentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ApartmentPage"
      },
      "ApartmentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "house_id": {
            "type": "integer",
            "title": "House Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "entrance": {
            "type": "integer",
            "title": "Entrance"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          }
        },
        "type": "object",
        "required": [
          "id",
          "house_id",
          "number",
          "entrance",
          "floor"
        ],
        "title": "ApartmentResponse"
      },
...
      "ContactInput": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Label"
          },
          "kind": {
            "type": "string",
            "enum": [
              "phone",
              "email",
              "address",
              "website"
            ],
            "title": "Kind"
          },
          "value": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Value"
          },
          "note": {
            "type": "string",
            "maxLength": 200,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "kind",
          "value"
        ],
        "title": "ContactInput"
      },
...
      "ContactsOutput": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContactInput"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "items"
        ],
        "title": "ContactsOutput"
      },
      "CreateIssue": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description"
          },
          "category": {
            "$ref": "#/components/schemas/IssueCategory"
          },
          "place": {
            "$ref": "#/components/schemas/Location"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "description",
          "category",
          "place",
          "request_id"
        ],
        "title": "CreateIssue"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HouseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "entrances_count": {
            "type": "integer",
            "title": "Entrances Count"
          },
          "floors_count": {
            "type": "integer",
            "title": "Floors Count"
          },
          "apartments_per_floor": {
            "type": "integer",
            "title": "Apartments Per Floor"
          }
        },
        "type": "object",
        "required": [
          "id",
          "address",
          "entrances_count",
          "floors_count",
          "apartments_per_floor"
        ],
        "title": "HouseResponse"
      },
      "IssueCategory": {
        "type": "string",
        "enum": [
          "water",
          "heating",
          "electricity",
          "elevator",
          "entrance",
          "yard",
          "garbage",
          "security",
          "other"
        ],
        "title": "IssueCategory"
      },
...
      "IssueHistory": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "at"
        ],
        "title": "IssueHistory"
      },
      "IssuePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "IssuePage"
      },
...
      "IssueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "house_id": {
            "type": "integer",
            "title": "House Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "$ref": "#/components/schemas/IssueCategory"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "mine": {
            "type": "boolean",
            "title": "Mine"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/IssueHistory"
            },
            "type": "array",
            "title": "History"
          },
          "photo_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Photo Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "house_id",
          "title",
          "description",
          "category",
          "status",
          "priority",
          "created_at",
          "mine",
          "address",
          "place",
          "history"
        ],
        "title": "IssueResponse"
      },
...
      "IssueStatus": {
        "type": "string",
        "enum": [
          "new",
          "in_progress",
          "resolved"
        ],
        "title": "IssueStatus"
      },
...
      "Location": {
        "properties": {
          "entrance": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Entrance"
          },
          "floor": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Floor"
          },
          "zone": {
            "type": "string",
            "enum": [
              "apartment",
              "corridor",
              "stairs",
              "elevator",
              "technical",
              "entrance",
              "house",
              "courtyard",
              "parking"
            ],
            "title": "Zone"
          },
          "apartment_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Apartment Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entrance",
          "floor",
          "zone"
        ],
        "title": "Location"
      },
...
      "Profile": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "apartments": {
            "items": {
              "$ref": "#/components/schemas/ApartmentResponse"
            },
            "type": "array",
            "title": "Apartments"
          },
          "houses": {
            "items": {
              "$ref": "#/components/schemas/HouseResponse"
            },
            "type": "array",
            "title": "Houses"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "apartments",
          "houses"
        ],
        "title": "Profile"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "MAXLaunchData": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}